refactor code for better visibility
